chore(api_apps): rename folder to models and add comments for app model
